13th commit, add checkout patient function

+++ src/main/java/checkout_patient.java
@@ -27,10 +27,10 @@ public class checkout_patient extends HttpServlet {
                 "<h2>Check out Patient</h2>\n" +
                 "<form action=\"checkout_patient\" method='post'>\n" +
                 "    Patient's first name:<br>\n" +
-                "    <input type=\"text\" name=\"firstname\" value=\"Enter patient's firstname \">\n" +
+                "    <input type=\"text\" name=\"firstname\" placeholder=\"Enter patient's firstname\" required>\n" +
                 "    <br><br>\n" +
                 "    Patient's last name:<br>\n" +
-                "    <input type=\"text\" name=\"lastname\" value=\"Enter patient's lastname\">\n" +
+                "    <input type=\"text\" name=\"lastname\" placeholder=\"Enter patient's lastname\" required>\n" +
                 "    <br><br>\n" +
                 "    <input type=\"submit\" value=\"Check out\">\n" +
                 "</form>\n" +
@@ -76,19 +76,21 @@ public class checkout_patient extends HttpServlet {
         } catch (SQLException e){
             e.printStackTrace();
         }
-        for (int i=0; i<c_r_p_fn.size(); i++) {
-            if (checkout_p_fn.equals(c_r_p_fn.get(i)) && checkout_p_ln.equals(c_r_p_ln.get(i))) {
-                patient_to_doctor pd = new patient_to_doctor(checkout_p_fn, checkout_p_ln, d_firstname, d_lastname);
-                response.getWriter().write(gson.toJson(pd));
-                try{
-                    Connection conn=DriverManager.getConnection(dbUrl);
-                    Statement s=conn.createStatement();
-                    String sqlcom="delete from patient_to_doctor_table where patient_firstname='"+checkout_p_fn+"' and patient_lastname='"+checkout_p_ln+"';";
-                    s.execute(sqlcom);
-                    conn.close();
-                    s.close();
-                }catch (SQLException e){
-                    e.printStackTrace();
+        if (checkout_p_fn!=null && checkout_p_ln!=null) {
+            for (int i = 0; i < c_r_p_fn.size(); i++) {
+                if (checkout_p_fn.equals(c_r_p_fn.get(i)) && checkout_p_ln.equals(c_r_p_ln.get(i))) {
+                    patient_to_doctor pd = new patient_to_doctor(checkout_p_fn, checkout_p_ln, d_firstname, d_lastname);
+                    response.getWriter().write(gson.toJson(pd));
+                    try {
+                        Connection conn = DriverManager.getConnection(dbUrl);
+                        Statement s = conn.createStatement();
+                        String sqlcom = "delete from patient_to_doctor_table where patient_firstname='" + checkout_p_fn + "' and patient_lastname='" + checkout_p_ln + "';";
+                        s.execute(sqlcom);
+                        conn.close();
+                        s.close();
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
                 }
             }
         }
